+++ data/files/products.json
@@ -1,0 +1,162 @@
+[
+  {
+    "id": "d3184ec2096974c5bf33f1b5",
+    "title": "Camisa de algodón a rayas",
+    "photo": "/camisa-rayas.jpg",
+    "category": "Ropa",
+    "price": 29.99,
+    "stock": 50
+  },
+  {
+    "id": "db2f75b49defaa1060b94300",
+    "title": "Zapatillas deportivas para correr",
+    "photo": "/zapatillas-corredor.jpg",
+    "category": "Calzado",
+    "price": 49.99,
+    "stock": 40
+  },
+  {
+    "id": "d5c0826c2b6bfc82bbf02597",
+    "title": "Mochila resistente al agua",
+    "photo": "/mochila-resistente.jpg",
+    "category": "Accesorios",
+    "price": 39.99,
+    "stock": 30
+  },
+  {
+    "id": "219027b68c72aae9d1f5edab",
+    "title": "Reloj de pulsera elegante",
+    "photo": "/reloj-elegante.jpg",
+    "category": "Accesorios",
+    "price": 69.99,
+    "stock": 25
+  },
+  {
+    "id": "b774a182b62b0304ff249fa5",
+    "title": "Auriculares inalámbricos Bluetooth",
+    "photo": "/auriculares-bluetooth.jpg",
+    "category": "Electrónica",
+    "price": 59.99,
+    "stock": 35
+  },
+  {
+    "id": "2f553efe3f9b4171f11f6681",
+    "title": "Pantalones vaqueros ajustados",
+    "photo": "/pantalones-vaqueros.jpg",
+    "category": "Ropa",
+    "price": 34.99,
+    "stock": 45
+  },
+  {
+    "id": "95b97106d6930b74c7b690d9",
+    "title": "Botella de agua deportiva",
+    "photo": "/botella-deportiva.jpg",
+    "category": "Accesorios",
+    "price": 14.99,
+    "stock": 55
+  },
+  {
+    "id": "ca44655bf98db71b9e6ad5d7",
+    "title": "Chaqueta ligera impermeable",
+    "photo": "/chaqueta-impermeable.jpg",
+    "category": "Ropa",
+    "price": 79.99,
+    "stock": 20
+  },
+  {
+    "id": "bee66e2ded3351eb9690940c",
+    "title": "Mochila para portátil",
+    "photo": "/mochila-portatil.jpg",
+    "category": "Accesorios",
+    "price": 49.99,
+    "stock": 30
+  },
+  {
+    "id": "379f9f719ea8cc5cb0b2d77c",
+    "title": "Teléfono inteligente de última generación",
+    "photo": "/telefono-inteligente.jpg",
+    "category": "Electrónica",
+    "price": 699.99,
+    "stock": 15
+  },
+  {
+    "id": "9ffe9fc1428f846095315ef7",
+    "title": "Gorra de béisbol ajustable",
+    "photo": "/gorra-beisbol.jpg",
+    "category": "Accesorios",
+    "price": 19.99,
+    "stock": 40
+  },
+  {
+    "id": "89eebd14c794da076b070e85",
+    "title": "Sudadera con capucha unisex",
+    "photo": "/sudadera-capucha.jpg",
+    "category": "Ropa",
+    "price": 44.99,
+    "stock": 25
+  },
+  {
+    "id": "aa97addfd8c8e872fd62cafe",
+    "title": "Bolsa de viaje resistente",
+    "photo": "/bolsa-viaje.jpg",
+    "category": "Accesorios",
+    "price": 59.99,
+    "stock": 35
+  },
+  {
+    "id": "e4da6f0e645c2aaba16e8875",
+    "title": "Bufanda de lana suave",
+    "photo": "/bufanda-lana.jpg",
+    "category": "Ropa",
+    "price": 24.99,
+    "stock": 50
+  },
+  {
+    "id": "1adaf3158ce5ac41b5ad7f41",
+    "title": "Botines de cuero elegantes",
+    "photo": "/botines-cuero.jpg",
+    "category": "Calzado",
+    "price": 89.99,
+    "stock": 20
+  },
+  {
+    "id": "eb4816fc415e8314de1803cf",
+    "title": "Cámara digital compacta",
+    "photo": "/img/camara-digital.jpg",
+    "category": "Electrónica",
+    "price": 199.99,
+    "stock": 10
+  },
+  {
+    "id": "37f47b2fb1b1b03ed9aae0f9",
+    "title": "Cinturón de cuero genuino",
+    "photo": "/img/cinturon-cuero.jpg",
+    "category": "Accesorios",
+    "price": 29.99,
+    "stock": 45
+  },
+  {
+    "id": "414d4e8cfd340043b24a9d1c",
+    "title": "Camiseta de algodón básica",
+    "photo": "/img/camiseta-algodon.jpg",
+    "category": "Ropa",
+    "price": 14.99,
+    "stock": 60
+  },
+  {
+    "id": "c0fab7529969921763c92793",
+    "title": "Maletín ejecutivo de cuero",
+    "photo": "/img/maletin-cuero.jpg",
+    "category": "Accesorios",
+    "price": 129.99,
+    "stock": 15
+  },
+  {
+    "id": "d2bcce95e0f5be58cf0c671e",
+    "title": "Gafas de sol polarizadas",
+    "photo": "/gafas-sol.jpg",
+    "category": "Accesorios",
+    "price": 39.99,
+    "stock": 30
+  }
+]
