+++ package.json
@@ -15,8 +15,5 @@
   },
   "keywords": [],
   "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "morgan": "^1.10.0"
-  }
+  "license": "ISC"
 }
