--- src/middlewares/errorHandler.js
@@ -1,8 +1,0 @@
-function errorHandler (error, req, res, next){
-    return res.json({
-        statusCode: error.statusCode || 500,
-        message: error.message || "ERROR"
-    })
-}
-
-export default errorHandler
