--- src/routers/index.router.js
@@ -1,8 +1,0 @@
-import { Router } from "express";
-import indexApiRouter from "./api/index.api.js";
-
-const indexRouter = Router()
-
-indexRouter.use("/api", indexApiRouter)
-
-export default indexRouter
